part_one_blog_code/requirements.txt
ollama
python-dotenv
langfuse
gradio 
httpx==0.27.2